swagger hecho

+++ src/main/java/com/api/books/services/impl/UserServiceImpl.java
@@ -4,6 +4,7 @@ import com.api.books.persistence.entities.UserEntity;
 import com.api.books.persistence.repositories.UserRepository;
 import com.api.books.services.UserService;
 import com.api.books.services.models.dtos.BookDTO;
+import com.api.books.services.models.dtos.UserRolesDTO;
 import com.api.books.services.models.dtos.templates.PasswordUpdateDTO;
 import com.api.books.services.models.dtos.templates.ResponseDTO;
 import com.api.books.services.models.dtos.UserDTO;
@@ -11,10 +12,15 @@ import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 @Service
@@ -34,6 +40,19 @@ public class UserServiceImpl implements UserService {
             UserEntity user = userRepository.findById(userId).orElse(null);
             if (user == null) return ResponseEntity.notFound().build();
             return ResponseEntity.ok(user.toDTO());
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
+    }
+
+    @Override
+    public ResponseEntity<UserRolesDTO> getUserRoles(Long userId) {
+        try {
+            if(!isADMIN())
+                return new ResponseEntity<>(new UserRolesDTO(), HttpStatus.UNAUTHORIZED);
+            UserEntity user = userRepository.findById(userId).orElse(null);
+            if (user == null) return ResponseEntity.notFound().build();
+            return ResponseEntity.ok(user.toRolesDTO());
         } catch (Exception e) {
             return ResponseEntity.internalServerError().build();
         }
@@ -67,8 +86,7 @@ public class UserServiceImpl implements UserService {
         }
     }
 
-    @Override
-    public ResponseEntity<ResponseDTO> update(Long id, UserEntity updatedUser) {
+    public ResponseEntity<ResponseDTO> updateUser(Long id, UserEntity updatedUser) {
         ResponseDTO response = new ResponseDTO();
         try {
             UserEntity previousUser = userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Usuario no encontrado"));
@@ -144,6 +162,8 @@ public class UserServiceImpl implements UserService {
     public ResponseEntity<ResponseDTO> removeUser(Long userId) {
         ResponseDTO response = new ResponseDTO();
         try {
+            if(!isADMIN())
+                return new ResponseEntity<>(new ResponseDTO("No tienes permiso"), HttpStatus.UNAUTHORIZED);
             userRepository.deleteById(userId);
             response.newMessage("Usuario borrado");
             return ResponseEntity.ok(response);
@@ -151,5 +171,19 @@ public class UserServiceImpl implements UserService {
             response.newError(e.toString());
             return ResponseEntity.internalServerError().build();
         }
+    }
+
+    @Override
+    public boolean isADMIN() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null && authentication.isAuthenticated()) {
+            Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
+            for (GrantedAuthority authority : authorities) {
+                if ("ADMIN".equals(authority.getAuthority())) {
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 }
